Move to go 1.20 error slice unwrapping
Build tag it to continue supporting older versions of go.

Fix a bug in the old Unwrap() method's implementation that would panic
if there were zero errors listed, due to an out of bounds index.
